# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: 后台服务接口
  version: 1.0.0
  title: BFFast快速开发平台
basePath: /fast
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
  /captcha.jpg:
    get:
      tags:
        - 登录登出
      summary: 获取验证码
      produces:
        - application/octet-stream
      parameters:
        - name: uuid
          description: UUID
          in: query
          type: string
          required: true
      responses:
        200:
          description: success
          schema:
            $ref: '#/definitions/Login'
  /sys/login:
    post:
      tags:
        - 登录登出
      summary: 登录
      produces:
        - application/json
      parameters:
        - name: body
          description: 登录信息
          in: body
          schema:
            $ref: '#/definitions/LoginDTO'
          required: true
      responses:
        200:
          description: success
          schema:
            $ref: '#/definitions/Login'
  /sys/login/social:
    post:
      tags:
        - 登录登出
      summary: 三方登录
      produces:
        - application/json
      parameters:
        - name: body
          description: 登录信息
          in: body
          schema:
            $ref: '#/definitions/SysLoginSocialDTO'
          required: true
      responses:
        200:
          description: success
          schema:
            $ref: '#/definitions/Login'
  /sys/login/mobile:
    post:
      tags:
        - 登录登出
      summary: 手机号登录
      produces:
        - application/json
      parameters:
        - name: body
          description: 登录信息
          in: body
          schema:
            $ref: '#/definitions/SysLoginMobileDTO'
          required: true
      responses:
        200:
          description: success
          schema:
            $ref: '#/definitions/Login'
...
  /sys/menu/list:
    get:
      tags:
        - 系统管理-菜单管理
      summary: 菜单列表
      produces:
        - application/json
      parameters:
        - name: clientType
          description: 客户端类型 0-运管中心 1-数据中心
          in: query
          type: string
          required: false
      responses:
        '200':
          description: 返回菜单列表
          schema:
            type: array
            items:
              $ref: '#/definitions/SysMenuEntity'
...
  /sys/menu/info/{menuId}:
    get:
      tags:
        - 系统管理-菜单管理
      summary: 获取菜单信息
      produces:
        - application/json
      parameters:
        - name: menuId
          description: 菜单ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回菜单信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              menu:
                description: 菜单信息
                $ref: '#/definitions/SysMenuEntity'
  /sys/menu/save:
    post:
      tags:
        - 系统管理-菜单管理
      summary: 添加菜单
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单对象
          in: body
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        200:
          description: success
          schema:
            $ref: '#/definitions/ResultDTO'
  /sys/menu/update:
    post:
      tags:
        - 系统管理-菜单管理
      summary: 修改菜单
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单对象
          in: body
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        200:
          description: success
          schema:
            $ref: '#/definitions/ResultDTO'
  /sys/menu/delete/{menuId}:
    post:
      tags:
        - 系统管理-菜单管理
      summary: 删除菜单
      produces:
        - application/json
      parameters:
        - name: menuId
          description: 菜单ID
          in: path
          type: integer
          required: true
      responses:
        200:
          description: success
          schema:
            $ref: '#/definitions/ResultDTO'
...
  /sys/menu/tree/{roleId}:
    get:
      tags:
        - 系统管理-菜单管理
      summary: 返回角色的菜单集合
      produces:
        - application/json
      parameters:
        - name: roleId
          description: 角色ID
          in: path
          type: integer
          required: true
      responses:
        200:
          description: success
          schema:
            $ref: '#/definitions/ResultDTO'
...
  /sys/dict/page:
    get:
      tags:
        - 系统管理-字典管理
      summary: 查看列表-分页
      produces:
        - application/json
      parameters:
        - name: current
          description: 页码
          in: query
          type: integer
          required: true
        - name: size
          description: 每页条数
          in: query
          type: integer
          required: true
      responses:
        200:
          description: 返回列表
          schema:
            $ref: '#/definitions/SysDictEntityListPage'
  #系统管理-字典管理-列表-不分页
...
  /sys/dict/type/{type}:
    get:
      tags:
        - 系统管理-字典管理
      summary: 通过类型查询字典列表
      produces:
        - application/json
      parameters:
        - name: type
          description: 字典类型code
          in: path
          type: string
          required: true
      responses:
        200:
          description: 返回列表
          schema:
            $ref: '#/definitions/SysDictEntityListPage'

  #系统管理-字典管理-保存
  /sys/dict/save:
    post:
      tags:
        - 系统管理-字典管理
      summary: 保存
      produces:
        - application/json
      parameters:
        - name: body
          description: SysDict 对象
          in: body
          schema:
            $ref: '#/definitions/SysDictEntity'
          required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/ResultDTO'

  #系统管理-字典管理-修改
  /sys/dict/update:
    post:
      tags:
        - 系统管理-字典管理
      summary: 修改
      produces:
        - application/json
      parameters:
        - name: body
          description: SysDict 对象
          in: body
          schema:
            $ref: '#/definitions/SysDictEntity'
          required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/ResultDTO'

  #系统管理-字典管理-单条查询
  /sys/dict/{id}:
    get:
      tags:
        - 系统管理-字典管理
      summary: 获取单条记录
      produces:
        - application/json
      parameters:
        - name: id
          description: 记录ID
          in: path
          type: integer
          required: true
      responses:
        200:
          description: 返回单条记录
          schema:
            type: object
            properties:
              code:
                description: 状态码 0-成功  非0-失败
                type: integer
                format: int32
              msg:
                description: 返回结果消息
                type: string
              data:
                $ref: '#/definitions/SysDictEntity'

  #系统管理-字典管理-删除
  /sys/dict/delete/{id}:
    post:
      tags:
        - 系统管理-字典管理
      summary: 删除
      produces:
        - application/json
      parameters:
        - name: id
          description: id
          in: path
          type: integer
          format: int64
          required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/ResultDTO'
  #系统管理-字典管理-列表-分页
  /sys/dict/type/page:
    get:
      tags:
        - 系统管理-字典管理
      summary: 字典类型分页列表
      produces:
        - application/json
      parameters:
        - name: current
          description: 页码
          in: query
          type: integer
          required: true
        - name: size
          description: 每页条数
          in: query
          type: integer
          required: true
      responses:
        200:
          description: 返回列表
          schema:
            $ref: '#/definitions/SysDictTypeEntityListPage'
  #系统管理-字典管理-列表-不分页
...
  /sys/dict/type/save:
    post:
      tags:
        - 系统管理-字典管理
      summary: 保存字典类型
      produces:
        - application/json
      parameters:
        - name: body
          description: SysDictType 对象
          in: body
          schema:
            $ref: '#/definitions/SysDictTypeEntity'
          required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/ResultDTO'
  #系统管理-字典管理-修改
  /sys/dict/type/update:
    post:
      tags:
        - 系统管理-字典管理
      summary: 修改字典类型
      produces:
        - application/json
      parameters:
        - name: body
          description: SysDictType 对象
          in: body
          schema:
            $ref: '#/definitions/SysDictTypeEntity'
          required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/ResultDTO'
  #系统管理-字典管理-删除
  /sys/dict/type/delete/{type}:
    post:
      tags:
        - 系统管理-字典管理
      summary: 删除字典类型
      produces:
        - application/json
      parameters:
        - name: type
          description: 类型编号
          in: path
          type: string
          required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/ResultDTO'
  #系统管理-版本管理-列表-分页
...
definitions:
  ResultDTO:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
        default: 0
      msg:
        description: 失败原因
        type: string
        default: success
      data:
        description: 数据对象
        type: object
...
  Login:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      msg:
        description: 失败原因
        type: string
      data:
        description: 数据对象
        type: object
        properties:
          token:
            description: token
            type: string
          expire:
            description: 过期时长
            type: integer
            format: int32
  LoginDTO:
    type: object
    properties:
      username:
        description: 用户名
        type: string
      password:
        description: 密码
        type: string
      captcha:
        description: 验证码
        type: string
      uuid:
        description: UUID
        type: string
      client:
        description: 客户端 admin-运管中心 data-数据中心 app-移动端
        type: string
  SysLoginSocialDTO:
    type: object
    properties:
      openId:
        description: openId
        type: string
      providerId:
        description: 三方平台 weixin qq weibo
        type: string
      accessToken:
        description: accessToken
        type: string
      refreshToken:
        description: refreshToken
        type: string
      expireTime:
        description: expireTime
        type: string
  SysLoginMobileDTO:
    type: object
    properties:
      mobile:
        description: 手机号
        type: string
      appKey:
        description: appKey
        type: string
      zone:
        description: zone
        type: string
      smsCode:
        description: 验证码
        type: string
...
  SysMenuEntity:
    type: object
    properties:
      menuId:
        description: 菜单ID
        type: integer
        format: int64
      name:
        description: 菜单名称
        type: string
      parentId:
        description: 父菜单ID，一级菜单为0
        type: integer
        format: int64
      parentName:
        description: 父菜单名称
        type: string
      path:
        description: 菜单URL
        type: string
      component:
        description: VUE页面
        type: string
      permission:
        description: 菜单权限标识
        type: string
      type:
        description: 类型 0-目录 1-菜单 2-按钮
        type: string
      clientType:
        description: 客户端类型 0-运管中心 1-数据中心
        type: string
      isSys:
        description: 是否系统菜单  0-否 1-是
        type: string
      icon:
        description: 菜单图标
        type: string
      sort:
        description: 排序
        type: integer
        format: int32
  SysMenuEntityEdit:
    type: object
    properties:
      menuId:
        description: 菜单ID
        type: integer
        format: int64
      name:
        description: 菜单名称
        type: string
      parentId:
        description: 父菜单ID，一级菜单为0
        type: integer
        format: int64
      path:
        description: 菜单URL
        type: string
      component:
        description: VUE页面
        type: string
      permission:
        description: 授权标识
        type: string
      type:
        description: 类型 0-目录 1-菜单 2-按钮
        type: string
      clientType:
        description: 客户端类型 0-运管中心 1-数据中心
        type: string
      isSys:
        description: 是否系统菜单  0-否 1-是
        type: string
      icon:
        description: 菜单图标
        type: string
      sort:
        description: 排序
        type: integer
        format: int32
...
  SysDictEntityListPage:
    type: object
    properties:
      code:
        description: 状态码 0-成功  非0-失败
        type: integer
        format: int32
      msg:
        description: 返回结果消息
        type: string
      data:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysDictEntity'
  #字典表-EntityList
...
  SysDictEntity:
    type: object
    properties:
      id:
        description: 编号
        type: integer
        format: int32
      value:
        description: 数据值
        type: string
      label:
        description: 标签名
        type: string
      type:
        description: 类型
        type: string
      description:
        description: 描述
        type: string
      sort:
        description: 排序（升序）
        type: integer
        format: int32
      createTime:
        description: 创建时间
        type: string
        format: date
      updateTime:
        description: 更新时间
        type: string
        format: date
      remarks:
        description: 备注信息
        type: string
      delFlag:
        description: 删除标记
        type: string
  #系统字典类型表-EntityListPage
  SysDictTypeEntityListPage:
    type: object
    properties:
      code:
        description: 状态码 0-成功  非0-失败
        type: integer
        format: int32
      msg:
        description: 返回结果消息
        type: string
      data:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysDictTypeEntity'
  #系统字典类型表-EntityList
...
  SysDictTypeEntity:
    type: object
    properties:
      typeValue:
        description: 字典类型值
        type: string
      typeName:
        description: 字典类型名
        type: string
  #系统管理-版本管理-EntityListPage
...